+++ packages/eyes-browser-extension/test/coverage/config.js
@@ -1,9 +1,11 @@
+const path = require('path')
+
 module.exports = {
   extends: 'https://raw.githubusercontent.com/applitools/sdk.coverage.tests/master/js/config.js',
   env: {
     NO_SDK: true,
-    SPEC_DRIVER_PATH: './test/utils/spec-driver.js',
-    SETUP_EYES_PATH: './test/utils/setup-eyes.js',
+    SPEC_DRIVER: path.resolve('./test/utils/spec-driver.js'),
+    SETUP_EYES: path.resolve('./test/utils/setup-eyes.js'),
   },
   overrides: [
     'https://raw.githubusercontent.com/applitools/sdk.coverage.tests/master/js/overrides.js',