+++ packages/eyes-sdk-core/lib/config/Configuration.js
@@ -65,6 +65,7 @@ const DEFAULT_VALUES = {
   // visual-grid
   concurrentSessions: 1,
   isThrowExceptionOn: false,
+  waitBeforeCapture: 100, //ms
 }
 
 /**
@@ -113,11 +114,12 @@ const DEFAULT_VALUES = {
 /**
  * @typedef PlainConfigurationClassic
  * @prop {boolean} forceFullPageScreenshot
- * @prop {number} waitBeforeScreenshots
+ * @prop {number} WaitBeforeScreenshots
  * @prop {StitchMode} stitchMode
  * @prop {number} stitchOverlap
  * @prop {boolean} hideScrollbars
  * @prop {boolean} hideCaret
+ * @prop {number} waitBeforeCapture
  */
 
 /**
@@ -222,6 +224,8 @@ class Configuration {
     this._hideCaret = undefined
     /** @type {number} */
     this._stitchOverlap = undefined
+    /** @type {number} */
+    this._waitBeforeCapture = undefined
 
     // visual grid
     /** @type {number} */
@@ -1058,6 +1062,13 @@ class Configuration {
   }
 
   /**
+   * @return {number} - The time to wait before each DOM capture and before the first screenshot in classic
+   */
+  getWaitBeforeCapture() {
+    return TypeUtils.getOrDefault(this._waitBeforeCapture, DEFAULT_VALUES.waitBeforeCapture)
+  }
+
+  /**
    * Sets the time to wait just before taking a screenshot (e.g., to allow positioning to stabilize when performing a
    * full page stitching).
    *
@@ -1070,6 +1081,22 @@ class Configuration {
       this._waitBeforeScreenshots = undefined
     } else {
       this._waitBeforeScreenshots = value
+    }
+    return this
+  }
+
+  /**
+   * Sets the time to wait just before each DOM capture and before the first screenshot captured in classic
+   *
+   * @param {number} value - The time to wait (Milliseconds). Values smaller or equal to 0, will cause the
+   *   default value to be used.
+   * @return {this}
+   */
+  setWaitBeforeCapture(value) {
+    if (value <= 0) {
+      this._waitBeforeCapture = undefined
+    } else {
+      this._waitBeforeCapture = value
     }
     return this
   }