+++ packages/eyes-sdk-core/lib/utils/takeDomSnapshots.js
@@ -14,9 +14,11 @@ async function takeDomSnapshots({
   getViewportSize,
   getEmulatedDevicesSizes,
   getIosDevicesSizes,
+  waitBeforeCapture,
 }) {
   if (!breakpoints) {
     logger.verbose(`taking single dom snapshot`)
+    await GeneralUtils.sleep(waitBeforeCapture)
     const snapshot = await takeDomSnapshot(logger, driver, {
       disableBrowserFetching,
       showLogs,
@@ -46,6 +48,7 @@ async function takeDomSnapshots({
   const snapshots = Array(browsers.length)
   if (requiredWidths.has(viewportSize.width)) {
     logger.log(`taking dom snapshot for existing width ${viewportSize.width}`)
+    await GeneralUtils.sleep(waitBeforeCapture)
     const snapshot = await takeDomSnapshot(logger, driver, {
       disableBrowserFetching,
       showLogs,
@@ -57,6 +60,7 @@ async function takeDomSnapshots({
     logger.log(`taking dom snapshot for width ${requiredWidth}`)
     try {
       await driver.setViewportSize({width: requiredWidth, height: viewportSize.height})
+      await GeneralUtils.sleep(waitBeforeCapture)
     } catch (err) {
       const actualViewportSize = await driver.getViewportSize()
       if (isStrictBreakpoints) {