+++ packages/spec-driver-selenium/src/index.ts
@@ -1,0 +1,7 @@
+if (!process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION) {
+  const {version} = require('selenium-webdriver/package.json')
+  const [major] = version.split('.', 1)
+  process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION = major
+}
+
+export * from './spec-driver'