+++ packages/eyes-selenium/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- move spec driver implementation to the separate package
 
 ## 4.56.3 - 2021/10/7